http: Non-Blocking IO
No longer: one thread == one connection
Darwin: uses kqueue https://developer.apple.com/library/archive/documentation/System/Conceptual/ManPages_iPhoneOS/man2/kqueue.2.html
Linux: uses io_uring https://man.archlinux.org/man/io_uring.7
Windows: uses a thread pool (IOCP in the future)

Downside is that retrieving body and a couple other functionalities are now using callbacks, this is because the IO abstraction is callback based using an event loop. API's and abstractions will be added to make this as pleasant to use as possible.

Middlewares now also have a papercut that needs to be designed out, when you call the next handler in your middleware you can no longer assume that the request is finished on the next line. So pre-handler middlewares are ok, but if you need to do cleanup it is not.